Moving deprecation_message from salt/utils/azurearm.py into salt/cloud/clouds/azurearm.py
