+++ salt/utils/azurearm.py
@@ -23,7 +23,6 @@ Azure (ARM) Utilities
 import importlib
 import logging
 import sys
-from functools import wraps
 from operator import itemgetter
 
 import salt.config
@@ -337,24 +336,3 @@ def compare_list_of_dicts(old, new, convert_id_to_name=None):
                 return ret
 
     return ret
-
-
-def deprecation_message(function):
-    """
-    Decorator wrapper to warn about azurearm deprecation
-    """
-
-    @wraps(function)
-    def wrapped(*args, **kwargs):
-        salt.utils.versions.warn_until(
-            "Chlorine",
-            "The 'azurearm' functionality in Salt has been deprecated and its "
-            "functionality will be removed in version 3007 in favor of the "
-            "saltext.azurerm Salt Extension. "
-            "(https://github.com/salt-extensions/saltext-azurerm)",
-            category=FutureWarning,
-        )
-        ret = function(*args, **salt.utils.args.clean_kwargs(**kwargs))
-        return ret
-
-    return wrapped
